% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_true.R
\name{bpr_true}
\alias{bpr_true}
\title{bpr_true}
\usage{
bpr_true(n, N, expected, theta_0, theta_1, p00, p11, p01, p10)
}
\arguments{
\item{n}{An integer specifying the number of tests that produced a particular result (either negative or positive) for the focal subject.}

\item{N}{Integer value giving the total number of repeated tests for the given subject.}

\item{expected}{An integer value (either 0 or 1) indicating whether the expected result is negative (0) or positive (1).}

\item{theta_0}{A real number (in range [0,1]) giving the prevalence of negative results for a set of test subjects.}

\item{theta_1}{A real number (in range [0,1]) giving the prevalence of positive results for a set of test subjects.}

\item{p00}{A real number (in range [0,1]) giving the probability that a test is negative given that the true result is negative (true-negative rate) for a set of test subjects.}

\item{p11}{A real number (in range [0,1]) giving the probability that a test is positive given that the true result is positive (true-positive rate) for a set of test subjects.}

\item{p01}{A real number (in range [0,1]) giving the probability that a test is negative given that the true result is positive (false-negative rate) for a set of test subjects.}

\item{p10}{A real number (in range [0,1]) giving the probability that a test is positive given that the true result is negative (false-positive rate) for a set of test subjects.}
}
\value{
A real number giving the Bayesian posterior probability that the test subject has a given result.
}
\description{
Calculates a Bayesian posterior probability that a given test subject (e.g. a gene) has a given expected result (either negative or positive).
}
