% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.perept.R
\name{EM.perept}
\alias{EM.perept}
\title{EM.perept}
\usage{
EM.perept(n_pos, N, delta_pos = NULL, maxiter = 1e+05, tol = 1e-06)
}
\arguments{
\item{n_pos}{Vector of integers specifying the number of tests that produced a positive result for each subject.}

\item{N}{Integer value giving the total number of repeated tests per subject.}

\item{delta_pos}{A vector of real numbers (in the range [0,1]) giving initial estimates of the probability of each subject being positive - must be the same length as `n_pos`. If `NULL`, then `n_pos`/`N` is used. Defaults to `NULL`.}

\item{maxiter}{A positive integer specifying the maximum allowable number of iterations without convergence occurring. Defaults to 1e5.}

\item{tol}{A real number specifying the maximum ODL difference between successive steps of the EM algorithm below which convergence occurs. Defaults to 1e-6.}
}
\value{
An object of class `perept`, which is a list with the following elements:
}
\description{
Expectation Maximisation (EM) algorithm to estimate false-positive and false-negative rates of a test when the true response is unknown and repeated tests are available for a set of test subjects (e.g. genes).
}
\references{
Jakobsdottir and Weeks (2007). Estimating prevalence, false-positive rate, and false-negative rate with use of repeated testing when true responses are unknown. Am J Hum Genet 81:1111-1113.
}
