% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPerf.R
\name{estPerf}
\alias{estPerf}
\title{estPerf}
\usage{
estPerf(n, N, delta)
}
\arguments{
\item{n}{Vector of integers specifying the number of tests that produced a particular result (either negative or positive) for each subject.}

\item{N}{Integer value giving the total number of repeated tests per subject.}

\item{delta}{Vector of real numbers giving the probability that a given gene has the expected result (either negative or positive) - must be the same length as `n`.}
}
\value{
The estimated performance, a real value in the range [0,1].
}
\description{
Estimate performance (false-positive rate, false-negative rate, true-positive rate, true-negative rate) when there are repeated tests for a set of test subjects (e.g. genes).
}
\examples{
\dontrun{estPerf(n,N,delta)}
}
\references{
Jakobsdottir and Weeks (2007). Estimating prevalence, false-positive rate, and false-negative rate with use of repeated testing when true responses are unknown. Am J Hum Genet 81:1111-1113.
}
