% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPrev.R
\name{estPrev}
\alias{estPrev}
\title{estPrev}
\usage{
estPrev(delta)
}
\arguments{
\item{delta}{Vector of real numbers giving the probability that a given gene has the expected result (either negative or positive).}
}
\value{
A real number giving the estimated prevalence of a given result (either negative or positive) for the set of subjects.
}
\description{
Estimate the prevalence of a particular result for a set of test subjects when there are repeated tests per subject (e.g. gene).
}
\examples{
\dontrun{estPrev(delta)}
}
