% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.perept.R
\name{MLE.perept}
\alias{MLE.perept}
\title{MLE.perept}
\usage{
MLE.perept(pos_counts, delta_pos = NULL, maxiter = 1e+05, tol = 1e-06)
}
\arguments{
\item{pos_counts}{A data frame of positive counts across a set of repeated tests with `n` columns. The first column is expected to contain subject IDs, columns 2 to `n-1` should contain positive counts across subjects for the `n-2` tests, and the `n`'th column should contain the total positive count across tests for each subject (the row sums).}

\item{delta_pos}{A vector of real numbers (in the range [0,1]) giving initial estimates of the probability of each subject being positive - must be the same length as `n_pos`. If `NULL`, then `n_pos`/`N` is used. Defaults to `NULL`.}

\item{maxiter}{A positive integer specifying the maximum allowable number of iterations without convergence occurring. Defaults to 1e5.}

\item{tol}{A real number specifying the maximum ODL difference between successive steps of the EM algorithm below which convergence occurs. Defaults to 1e-6.}
}
\value{
An object of class `MLE.perept`, which is a list with the following elements:
}
\description{
Maximum Likelihood Estimation (MLE) of the performance of a test when there are repeated tests performed on the same subjects.
}
\references{
Jakobsdottir and Weeks (2007). Estimating prevalence, false-positive rate, and false-negative rate with use of repeated testing when true responses are unknown. Am J Hum Genet 81:1111-1113.
}
