% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODL.R
\name{ODL}
\alias{ODL}
\title{ODL}
\usage{
ODL(n_pos, N, theta_0, theta_1, p00, p11, p01, p10)
}
\arguments{
\item{n_pos}{Vector of integers specifying the number of tests that produced a positive result for each subject.}

\item{N}{Integer value giving the total number of repeated tests per subject.}

\item{theta_0}{A real number (in range [0,1]) giving the prevalence of negative results for a set of test subjects.}

\item{theta_1}{A real number (in range [0,1]) giving the prevalence of positive results for a set of test subjects.}

\item{p00}{A real number (in range [0,1]) giving the probability that a test is negative given that the true result is negative (true-negative rate) for a set of test subjects.}

\item{p11}{A real number (in range [0,1]) giving the probability that a test is positive given that the true result is positive (true-positive rate) for a set of test subjects.}

\item{p01}{A real number (in range [0,1]) giving the probability that a test is negative given that the true result is positive (false-negative rate) for a set of test subjects.}

\item{p10}{A real number (in range [0,1]) giving the probability that a test is positive given that the true result is negative (false-positive rate) for a set of test subjects.}
}
\value{
A real value giving the `log(ODL)`.
}
\description{
Calculate the logarithm of the Observed Data Likelihood (ODL) when true responses are unknown and repeated tests are performed on a set of test subjects (e.g. genes).
}
