% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.perept.R
\name{print.EM.perept}
\alias{print.EM.perept}
\title{print.EM.perept}
\usage{
\method{print}{EM.perept}(x, ...)
}
\arguments{
\item{x}{An object of class `EM.perept`.}

\item{...}{Other arguments to be passed to print.}
}
\value{
Prints a summary to the console screen.
}
\description{
print S3 method for `EM.perept`` objects.
}
