% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.perept.R
\name{print.perept}
\alias{print.perept}
\title{print.perept}
\usage{
\method{print}{perept}(x, ...)
}
\arguments{
\item{x}{An object of class `perept`.}

\item{...}{Other arguments to be passed to print.}
}
\value{
Prints a summary to the console screen.
}
\description{
print S3 method for perept.
}
