% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wong_etal.n_pos.R
\docType{data}
\name{wong_etal.n_pos}
\alias{wong_etal.n_pos}
\title{wong_etal.n_pos}
\format{A vector of integers of length 24392.}
\usage{
data(wong_etal.n_pos)
}
\description{
Positive test counts for six CNV tests applied to the same set of 23k clones (Wong et al. (2007)).
}
\references{
Wong et al. (2007) A Comprehensive Analysis of Common Copy-Number Variations in the Human Genome. Am J Hum Genet 80:91-104.
}
\keyword{CNVs}
\keyword{repeated-tests}
